Implement core networking, logging, and bandwidth components
Replace stub implementations with full functionality:
- net/connection.cpp: TCP connection handling with async I/O
- net/acceptor.cpp: Connection acceptor implementation
- util/logging.cpp: Console and file logging sinks
- util/config.cpp: TOML configuration parsing
- policy/bandwidth.cpp: Bandwidth rate limiting

Updates to headers and tests to support the implementations.

+++ src/crypto/aes_ctr.cpp
@@ -92,7 +92,7 @@ AesCtr128::process(std::span<const uint8_t> data) {
         return std::unexpected(AesError::EncryptionFailed);
     }
 
-    output.resize(len);
+    output.resize(static_cast<size_t>(len));
     return output;
 }
